split script.js
