+++ gulpfile.js
@@ -29,7 +29,7 @@ gulp.task('css', function () {
          .pipe(browserSync.stream());
 });
 
-gulp.task('browserify', function(){
+gulp.task('script', function(){
  var bundleStream = browserify('./js/script.js').bundle()
 
   return bundleStream
@@ -40,7 +40,7 @@ gulp.task('browserify', function(){
     .pipe(browserSync.stream());
 });
 
-gulp.task('default', gulp.parallel('less', 'css', 'browserify', function(){
+gulp.task('default', gulp.parallel('less', 'css', 'script', function(){
        
        browserSync.init({
        	server: {baseDir: './'}
@@ -48,7 +48,7 @@ gulp.task('default', gulp.parallel('less', 'css', 'browserify', function(){
 
     gulp.watch('./index.html').on('change', browserSync.reload);
     gulp.watch('./style/main.css').on('change', browserSync.reload);
-    gulp.watch('./js/script.js').on('change', gulp.series('browserify'), browserSync.reload);
+    gulp.watch('./js/script.js').on('change', gulp.series('script'), browserSync.reload);
     gulp.watch('./style/less/*.less', gulp.series('less'));
     gulp.watch('./style/dest/*.css', gulp.series('css'));
 }));
