+++ package.json
@@ -26,6 +26,7 @@
     "gulp-less": "^4.0.1",
     "gulp-order": "^1.2.0",
     "gulp-rename": "^1.4.0",
+    "minimatch": "^3.0.4",
     "normalize.css": "^8.0.1",
     "npm": "^6.9.0",
     "validate-js": "github:rickharrison/validate.js#master",
